Add prometheus middleware to cargohold
