+++ services/cargohold/package.yaml
@@ -36,6 +36,7 @@ library:
   exposed-modules:
   - CargoHold.API
   - CargoHold.Options
+  - CargoHold.Run
   dependencies:
   - base >=4 && <5
   - attoparsec >=0.12
