+++ services/cargohold/src/Main.hs
@@ -1,7 +1,7 @@
 module Main (main) where
 
 import Imports
-import CargoHold.API
+import CargoHold.Run (run)
 import OpenSSL (withOpenSSL)
 
 import Util.Options
@@ -9,7 +9,7 @@ import Util.Options
 main :: IO ()
 main = withOpenSSL $ do
     options <- getOptions desc Nothing defaultPath
-    runServer options
+    run options
   where
     desc = "Cargohold - Asset Storage"
     defaultPath = "/etc/wire/cargohold/conf/cargohold.yaml"
